requirements.txt
crawl4ai
torch
sentence-transformers
faiss-cpu
numpy
transformers==4.35.2
huggingface-hub
requests
urllib3
idna
certifi
tqdm
pyyaml
filelock
packaging
regex
typing_extensions
safetensors
scikit-learn
scipy
joblib
threadpoolctl
tokenizers>=0.14,<0.19
pillow